+++ deno_webgpu/lib.rs
@@ -406,7 +406,7 @@ pub async fn op_webgpu_request_adapter(
             wgpu_types::InstanceDescriptor {
                 backends,
                 dx12_shader_compiler: wgpu_types::Dx12Compiler::Fxc,
-                force_angle_gles31: false
+                gles_minor_version: wgpu_types::Gles3MinorVersion::default(),
             },
         )));
         state.borrow::<Instance>()